o the components now grab the mouse, so the mouse can move out of the component and still change the values o fixed some resizing and settings loading bugs in colour patches o in triangle selector, switched from qreal to int. int seems to have a better handling of bounds etc.
svn path=/trunk/koffice/; revision=1156116

+++ krita/plugins/extensions/dockers/colorselectorng/kis_color_selector_triangle.h
@@ -35,12 +35,12 @@ protected:
     void paint(QPainter*);
     bool isComponent(int x, int y) const;
 private:
-    qreal triangleWidth() const;
-    qreal triangleHeight() const;
+    int triangleWidth() const;
+    int triangleHeight() const;
     void updatePixelCache();
     QColor colorAt(int x, int y) const;
     QPoint widgetToTriangleCoordinates(const QPoint& point) const;
-    QPoint triangleToWidgetCoordinates(const QPoint &point) const;
+    QPoint triangleToWidgetCoordinates(const QPoint& point) const;
     QImage m_pixelCache;
     
     QPointF m_lastClickPos;
